Refactor: 채팅 패키지 구조 변경
