+++ src/main/java/com/f3/exercise_mate/chat/repository/FakeChatRoomRepository.java
@@ -1,5 +1,6 @@
 package com.f3.exercise_mate.chat.repository;
 
+import com.f3.exercise_mate.chat.application.interfaces.ChatRoomRepository;
 import com.f3.exercise_mate.chat.domain.ChatRoom;
 import com.f3.exercise_mate.common.exception.ErrorCode;
 import com.f3.exercise_mate.common.exception.ExerciseMateException;
